archiver/requirements.txt
PyYAML>=6.0
schedule>=1.2.0
docker>=7.0.0
